% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url2file.R
\name{url2file}
\alias{url2file}
\title{url2file
This function downloads a file from a web address and places it in a file in your working directory.
Arguments:
urlFile: provide a url address pointing to the file to download stored in a vector
dirID: name of the directory to save files in (default = "Downloaded_files");}
\usage{
url2file(urlFile, dirID = "Downloaded_files")
}
\description{
url2file
This function downloads a file from a web address and places it in a file in your working directory.
Arguments:
urlFile: provide a url address pointing to the file to download stored in a vector
dirID: name of the directory to save files in (default = "Downloaded_files");
}
