% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url2file.R
\name{url2file}
\alias{url2file}
\title{Download a file from a URL into your working directory}
\usage{
url2file(urlfile, dirID = "Downloaded_files")
}
\arguments{
\item{urlfile}{provide a URL adddress in quotes of the file to download stored in a vector}

\item{dirID}{provide the name of an existing or new folder (default = "Downloaded_files")

#' @examples
You can specify an existing folder in the working directory (dirID="myfolder").
Or specify a new folder (dirID="newfolder")}
}
\description{
This is a fuction to streamline downloading files from a URL to a new or existing folder in the working directory.
}
\examples{
Passing the address to "url2file("https://page.com")" will download the page.
This will be into a file called "Downloaded_files" in your working directory.
}
